Do not add spaces to prefix/suffix
Adding then removing a comment block left a trailing space on the line.
While I originally changed the code to remove that space, I decided
it was better to make as few changes as possible to the code and not
assume that we put a space there as we could be removing a comment
that was put there manually.